2023-07-18 webpack

+++ src/controllers/videoController.js
@@ -33,11 +33,17 @@ export const getEdit = async (req, res) => {
 };
 
 export const postEdit = async (req, res) => {
+	const {
+		user: { _id },
+	} = req.session;
 	const { id } = req.params;
 	const { title, description, hashtags } = req.body;
 	const video = await Video.exists({ _id: id });
 	if (!video) {
 		return res.status(404).render("404", { pageTitle: "Video not found." });
+	}
+	if (String(video.owner) !== String(_id)) {
+		return res.status(403).redirect("/");
 	}
 	await Video.findByIdAndUpdate(id, {
 		title,
@@ -82,7 +88,14 @@ export const postUpload = async (req, res) => {
 };
 
 export const deleteVideo = async (req, res) => {
+	const {
+		user: { _id },
+	} = req.session;
 	const { id } = req.params;
+	const video = await Video.findById(id);
+	if (String(video.owner) !== String(_id)) {
+		return res.status(403).redirect("/");
+	}
 	await Video.findByIdAndDelete(id);
 	return res.redirect("/");
 };
